further split the tests
